Decouple sending video and commands
* Enables choosing different parameters (such as socket type) for different message types.
